move cta into subcomponent
